# expression: implement vectorized evaluation for builtinMicroSecondSig