--- a/expression/builtin_time_vec.go
+++ b/expression/builtin_time_vec.go
@@ -601,7 +601,6 @@ func (b *builtinMicroSecondSig) vecEvalInt(input *chunk.Chunk, result *chunk.Col
 	return nil
 }
 
-
 func (b *builtinSubDatetimeAndStringSig) vectorized() bool {
 	return false
 }