### What problem does this PR solve? <!--add issue link with summary if exists-->
expression: implement vectorized evaluation for `builtinMicroSecondSig` for #12103.

### What is changed and how it works?
```
goos: linux
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinTimeFuncGenerated-4   	1000000000	         0.00774 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinMicroSecondSig-VecBuiltinFunc-4         	  139454	      8298 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinMicroSecondSig-NonVecBuiltinFunc-4      	   43777	     27107 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/pingcap/tidb/expression	2.785s
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test