@francis0407 2019-11-15
The conflicts are not resolved yet. Run `git pull upstream master` to merge the master into your branch and resolve the conflicts.